requirements.txt
psycopg2
pyyaml
uszipcode

# Dev only?
pandas
csvkit